Create #80 Removing Duplicates from Sorted Array II
First, we are going to create two pointers for a counter and an itr. We do a range from 1 to len(nums). Then we look at the spot behind the one we're at right now. If it is there, we increment the counter or else we set it back to 1. If it 2 or less, we place the i in the itr. Then we move up the itr. Return the itr to know how many we have correct.