Create #2095 Delete the Middle Node of a Linked List
This one is super easy, we just do the fast slow pointers except we need to make sure to start the fast at 2 up and then check if fast and fast.next in the while condition. This makes it so that when we break, we can just skip one and call it good. Don't forget that you also need to make sure to check if head.next is None at the start.